Create Stack_using_Array.java
Write a program to implement the two stacks using one array. Consider the size of an array as 1000.



Format:

Input:

Get the number of elements to be inserted in stack 1.
And then, get the elements one by one.
Also, get the number of elements to be inserted in stack 2.
And then, get the elements one by one.
Get the number of elements to be deleted from stack 1.
Get the number of elements to be deleted from stack 2.
Output:

Print "Stack underflow. pop from stack 1 failed" message if the pop operation fails while doing it in stack 1.
Print the elements of stack 1.
Print "Stack underflow. pop from stack 1 failed" message if the pop operation fails while doing it in stack 2.
Print the elements of stack 2.